refactor: move crash landing to its own package (#1050)
