+++ app/src/main/java/com/eddyizm/tempus/ui/crash/CrashLogsFragment.java
@@ -1,16 +1,11 @@
-package com.eddyizm.tempus.ui.fragment;
+package com.eddyizm.tempus.ui.crash;
 
-import android.content.ClipData;
-import android.content.ClipboardManager;
-import android.content.Context;
-import android.content.Intent;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -19,7 +14,6 @@ import androidx.fragment.app.Fragment;
 import androidx.media3.common.util.UnstableApi;
 
 import com.eddyizm.tempus.R;
-import com.eddyizm.tempus.ui.activity.CrashActivity;
 
 public class CrashLogsFragment extends Fragment {
